+++ microcontroller_morse/src/morse_utils.c
@@ -10,6 +10,7 @@ extern bool  tim3_done;
 extern bool  tim4_done;
 extern Pin_t led1;
 
+/// @brief Waits a short amount of time
 void short_wait() {
     start_tim(TIM2);
     while (!tim2_done) {
@@ -19,6 +20,7 @@ void short_wait() {
     stop_tim(TIM2);
 }
 
+/// @brief Waits a medium amount of time
 void medium_wait() {
     start_tim(TIM3);
     while (!tim3_done) {
@@ -27,6 +29,7 @@ void medium_wait() {
     stop_tim(TIM3);
 }
 
+/// @brief Waits a long amount of time
 void long_wait() {
     start_tim(TIM4);
     while (!tim4_done) {
@@ -36,12 +39,14 @@ void long_wait() {
     stop_tim(TIM4);
 }
 
+/// @brief Makes the LED turn on for a short amount of time
 void short_blink() {
     set_pin(&led1);
     short_wait();
     clear_pin(&led1);
 }
 
+/// @brief Makes the LED turn on for a medium amount of time
 void medium_blink() {
     set_pin(&led1);
     medium_wait();
@@ -55,6 +60,9 @@ char *morse_letters[] = {"---",  "-...", "-.-.", "-..",  ".",    "..-.", "--.",
 char *morse_digits[]  = {"-----", ".----", "..---", "...--", "....-",
                          ".....", "-....", "--...", "---..", "----."};
 
+/// @brief Converts a uint8_t punctuation into its morse version
+/// @param c The punctuation to convert
+/// @return A string of morse code
 char *get_morse_punct(uint8_t c) {
     switch (c) {
         case '.':
@@ -117,12 +125,14 @@ char *get_morse_punct(uint8_t c) {
     }
 }
 
-void string_to_morse(uint8_t *str, int str_length) {
+void string_to_morse(uint8_t *str, uint8_t str_length) {
     for (int i = 0; i < str_length; i++) {
         uint8_t current_char = str[i];
+        /* for each character in the string */
         if (current_char == ' ') {
             long_wait();
         } else {
+            /* converts to morse the character */
             char *morse_code;
             if (isalpha(current_char)) {
                 morse_code = morse_letters[current_char - 'a'];
@@ -132,6 +142,7 @@ void string_to_morse(uint8_t *str, int str_length) {
                 morse_code = get_morse_punct(current_char);
             }
             for (int j = 0; j < strlen(morse_code); j++) {
+                /* for dot/dash in the morse representation */
                 char current_morse_char = morse_code[j];
                 switch (current_morse_char) {
                     case '.':
@@ -141,9 +152,11 @@ void string_to_morse(uint8_t *str, int str_length) {
                     default:
                         medium_blink();
                 }
+                /* between each dot/dash waits a short amount of time */
                 short_wait();
             }
         }
+        /* between each letter waits a medium amount of time */
         medium_wait();
     }
 }
